refactor(e2e): relocate prover tests to single-node
Move e2e_prover/{full,client}.test.ts into single-node/prover/ and drop
the redundant describe-title prefixes (full_prover/client_prover ->
single-node/prover/{full,client}). Repoint the two special CI jobs for
the heavy real/fake full-prover run at the new path, preserving the
CI_FULL real (CPUS=16:MEM=96g:TIMEOUT=20m, e2e_prover_full_real) vs fake
(FAKE_PROOFS=1, e2e_prover_full_fake) split. The prover/ subfolder is
intentionally left out of the generic test glob, matching today: full
runs only via its dedicated jobs and client runs in no CI job.
